+++ backend/src/main/java/com/dubu/backend/plan/api/response/FeedbackWritePageInfoResponse.java
@@ -1,7 +1,7 @@
 package com.dubu.backend.plan.api.response;
 
 import com.dubu.backend.plan.domain.Plan;
-import com.dubu.backend.todo.domain.Todo;
+import com.dubu.backend.plan.domain.Todo;
 
 import java.util.List;
