+++ backend/src/main/java/com/dubu/backend/plan/api/response/RecentPlanTodosResponse.java
@@ -1,6 +1,6 @@
 package com.dubu.backend.plan.api.response;
 
-import com.dubu.backend.todo.domain.Todo;
+import com.dubu.backend.plan.domain.Todo;
 
 import java.util.List;
